+++ lib/xcodeproj/project.rb
@@ -1,5 +1,6 @@
 require 'fileutils'
 require 'pathname'
+require 'securerandom'
 
 require 'xcodeproj/project/object'
 require 'xcodeproj/project/project_helper'
@@ -169,7 +170,7 @@ module Xcodeproj
     #
     def initialize_from_file
       pbxproj_path = path + 'project.pbxproj'
-      plist = Xcodeproj.read_plist(pbxproj_path.to_s)
+      plist = Xcodeproj::PlistHelper.read(pbxproj_path.to_s)
       root_object_uuid = plist['rootObject']
       root_object.remove_referrer(self) if root_object
       @root_object = new_from_plist(root_object_uuid, plist['objects'], self)
@@ -300,7 +301,7 @@ module Xcodeproj
       save_path ||= path
       FileUtils.mkdir_p(save_path)
       file = File.join(save_path, 'project.pbxproj')
-      Xcodeproj.write_plist(to_hash, file)
+      Xcodeproj::PlistHelper.write(to_hash, file)
       fix_encoding(file)
       XCProjHelper.touch(save_path) unless disable_xcproj
     end
@@ -404,7 +405,7 @@ module Xcodeproj
     # @return [void]
     #
     def generate_available_uuid_list(count = 100)
-      new_uuids = (0..count).map { Xcodeproj.generate_uuid }
+      new_uuids = (0..count).map { SecureRandom.hex(12).upcase }
       uniques = (new_uuids - (@generated_uuids + uuids))
       @generated_uuids += uniques
       @available_uuids += uniques
@@ -480,7 +481,7 @@ module Xcodeproj
     #
     def reference_for_path(absolute_path)
       absolute_pathname = Pathname.new(absolute_path)
-      
+
       unless absolute_pathname.absolute?
         raise ArgumentError, "Paths must be absolute #{absolute_path}"
       end
@@ -702,7 +703,7 @@ module Xcodeproj
       end
 
       xcschememanagement_path = schemes_dir + 'xcschememanagement.plist'
-      Xcodeproj.write_plist(xcschememanagement, xcschememanagement_path)
+      Xcodeproj::PlistHelper.write(xcschememanagement, xcschememanagement_path)
     end
 
     #-------------------------------------------------------------------------#